fix: resolve remaining clippy errors and Windows LLVM setup
- Add unsafe blocks at call sites for append_element/set_element_string (worker.rs)
- Remove unnecessary u64 casts in alloc_profile.rs
- Use &Path instead of &PathBuf in live_subscription.rs
- Install LLVM 18 on Windows via install-llvm-action for bindgen
- Remove hardcoded LIBCLANG_PATH (both actions set it automatically)

+++ crates/xbbg-bench/benches/alloc_profile.rs
@@ -90,8 +90,8 @@ where
 
     AllocStats {
         name: name.to_string(),
-        bytes_allocated: bytes as u64,
-        allocations: allocs as u64,
+        bytes_allocated: bytes,
+        allocations: allocs,
         duration_us: duration.as_micros() as u64,
         bytes_per_op: bytes as f64 / iterations as f64,
         allocs_per_op: allocs as f64 / iterations as f64,
